--- birthdayFunding/birthdayFunding/src/main/java/com/just/birthdayFunding/dto/LoginResponse.java
@@ -1,11 +1,0 @@
-package com.just.birthdayFunding.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class LoginResponse {
-    private String accessToken;
-    private String refreshToken;
-}
