--- birthdayFunding/birthdayFunding/src/main/java/com/just/birthdayFunding/service/KakaoLoginService.java
@@ -1,159 +1,0 @@
-package com.just.birthdayFunding.service;
-
-import com.just.birthdayFunding.dto.LoginResponse;
-import com.nimbusds.jose.shaded.gson.JsonElement;
-import com.nimbusds.jose.shaded.gson.JsonParser;
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.OutputStreamWriter;
-import java.net.HttpURLConnection;
-import java.net.MalformedURLException;
-import java.net.ProtocolException;
-import java.net.URL;
-import java.util.HashMap;
-import java.util.Map;
-import net.minidev.json.JSONObject;
-import net.minidev.json.parser.JSONParser;
-import net.minidev.json.parser.ParseException;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-
-@Service
-public class KakaoLoginService {
-    String requestURL = "https://kauth.kakao.com/oauth/token";
-    @Value("${kakao.auth.client_id}")
-    private String client_id;
-
-    @Value("${kakao.auth.redirect_uri}")
-    private String redirect_uri;
-
-    public LoginResponse getKakaoTokens(String code) {
-        String accessToken = "";
-        String refreshToken = "";
-
-        try {
-            URL url = new URL(requestURL);
-            HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-
-            conn.setRequestMethod("POST");
-            // setDoOutput()은 OutputStream으로 POST 데이터를 넘겨 주겠다는 옵션이다.
-            // POST 요청을 수행하려면 setDoOutput()을 true로 설정한다.
-            conn.setDoOutput(true);
-
-            // POST 요청에서 필요한 파라미터를 OutputStream을 통해 전송
-            BufferedWriter bufferedWriter = new BufferedWriter(new OutputStreamWriter(conn.getOutputStream()));
-            StringBuilder sb = new StringBuilder();
-            sb.append("grant_type=authorization_code");
-            sb.append("&client_id=" + client_id);     // REST_API_KEY
-            sb.append("&redirect_uri=" + redirect_uri); // REDIRECT_URI
-            sb.append("&code=" + code);
-
-            bufferedWriter.write(sb.toString());
-            bufferedWriter.flush();
-
-            int responseCode = conn.getResponseCode();
-
-            // 요청을 통해 얻은 데이터를 InputStreamReader을 통해 읽어 오기
-            BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(conn.getInputStream()));
-            String line = "";
-            StringBuilder result = new StringBuilder();
-
-            while ((line = bufferedReader.readLine()) != null) {
-                result.append(line);
-            }
-            JsonElement element = JsonParser.parseString(result.toString());
-
-            accessToken = element.getAsJsonObject().get("access_token").getAsString();
-            refreshToken = element.getAsJsonObject().get("refresh_token").getAsString();
-
-            bufferedReader.close();
-            bufferedWriter.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        return new LoginResponse(accessToken, refreshToken);
-    }
-
-    public Map<String, String> getUserInfo(String access_token) throws IOException {
-        String host = "https://kapi.kakao.com/v2/user/me";
-        Map<String, String> result = new HashMap<>();
-        try {
-            URL url = new URL(host);
-
-            HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
-            urlConnection.setRequestProperty("Authorization", "Bearer " + access_token);
-            urlConnection.setRequestMethod("GET");
-
-            int responseCode = urlConnection.getResponseCode();
-            System.out.println("responseCode = " + responseCode);
-
-            BufferedReader br = new BufferedReader(new InputStreamReader(urlConnection.getInputStream()));
-            String line = "";
-            String res = "";
-            while ((line = br.readLine()) != null) {
-                res += line;
-            }
-            br.close();
-            System.out.println("res = " + res);
-
-            JSONParser parser = new JSONParser();
-            JSONObject obj = (JSONObject) parser.parse(res);
-            JSONObject kakao_account = (JSONObject) obj.get("kakao_account");
-            JSONObject properties = (JSONObject) obj.get("properties");
-            System.out.println("d:d:d");
-            System.out.println(obj.toString());
-            String id = obj.getAsString("id");
-            String nickname = properties.getAsString("nickname");
-            String imagePath = properties.getAsString("profile_image");
-
-
-            System.out.println("d:d:d");
-            result.put("id", id);
-            result.put("nickname", nickname);
-            result.put("imagePath", imagePath);
-
-
-
-
-        } catch (IOException | ParseException e) {
-            e.printStackTrace();
-        }
-        System.out.println(result);
-        return result;
-    }
-
-    public String getAgreementInfo(String access_token) {
-        String result = "";
-        String host = "https://kapi.kakao.com/v2/user/scopes";
-        try {
-            URL url = new URL(host);
-            HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
-            urlConnection.setRequestMethod("GET");
-            urlConnection.setRequestProperty("Authorization", "Bearer " + access_token);
-
-            BufferedReader br = new BufferedReader(new InputStreamReader(urlConnection.getInputStream()));
-            String line = "";
-            while ((line = br.readLine()) != null) {
-                result += line;
-            }
-
-            int responseCode = urlConnection.getResponseCode();
-            System.out.println("responseCode = " + responseCode);
-
-            // result is json format
-            br.close();
-
-        } catch (MalformedURLException e) {
-            e.printStackTrace();
-        } catch (ProtocolException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return result;
-    }
-
-}
